Add support for unbounded ranges (#91)
* Add support for unbounded ranges

This commit adds support for open ranges, i.e. inequality operators in
front of a term. In tree form, the < is named To, and > is named From.

Additionally, this commit also adds a TreeTransformer to convert these
open ranges to more traditional Range objects.

To properly support escaping, this commit also adjusts how escaping
sequences work. After careful evaluation of how Apache Lucene handles
escape sequences, it appears that random characters can be escaped, even
if they result in unknown escape sequences: the escaped character is
always yielded. This makes support for operations such as `<\=foo` a lot
less complicated.

This commit does NOT add support in the ElasticsearchQueryBuilder.

Authored-by: JSCU-CNI
Co-authored-by: Alex Garel <alex@garel.org>